Added test-utils crate
  - Created a new test-utils crate, to contain common functions and data
    structures for the purpose of testing.

  - Moved common functionality that has a wider scope from
    storage::common to tests-utils::storage.
