+++ Cargo.toml
@@ -24,6 +24,7 @@ members = [
   "./crates/storage",
   "./crates/store",
   "./crates/store/blobs",
+  "./crates/test-utils",
 
   "./apps/kv-store",
   "./apps/only-peers",
