+++ crates/test-utils/src/lib.rs
@@ -1,0 +1,3 @@
+//! Test utilities and data structures for use in the Calimero crates.
+
+pub mod storage;
